changed files routing and controller file
